move condition to conditions/base
