+++ src/conditions/base/condition.ts
