+++ src/conditions/base/evm.ts
@@ -1,12 +1,13 @@
 import Joi from 'joi';
 
-import { Condition, makeReturnValueTest } from '../condition';
 import {
   ETH_ADDRESS_REGEXP,
   SUPPORTED_CHAINS,
   USER_ADDRESS_PARAM,
 } from '../const';
 import { ContextParametersHandlerMixin } from '../context/mixin';
+
+import { Condition, makeReturnValueTest } from './condition';
 
 export interface EvmConditionConfig {
   CONDITION_TYPE: string;
