+++ src/conditions/index.ts
@@ -2,7 +2,7 @@ import * as base from './base';
 import * as predefined from './predefined';
 
 export { predefined, base };
-export { Condition } from './condition';
+export { Condition } from './base/condition';
 export { ConditionSet } from './condition-set';
 export { Operator } from './operator';
 export type { CustomContextParam } from './context';
